refactor :: 의존성을 data 계층으로 통일
