+++ data/src/main/kotlin/team/aliens/dms/android/data/latestudy/mapper/StudyTypeMapper.kt
@@ -1,6 +1,6 @@
 package team.aliens.dms.android.data.latestudy.mapper
 
-import team.aliens.dms.android.domain.latestudy.model.StudyType
+import team.aliens.dms.android.data.latestudy.model.StudyType
 import team.aliens.dms.android.network.latestudy.model.FetchStudyTypesResponse
 
 fun FetchStudyTypesResponse.toModel(): List<StudyType> =
