+++ data/src/main/kotlin/team/aliens/dms/android/data/latestudy/model/StudyType.kt
@@ -1,0 +1,6 @@
+package team.aliens.dms.android.data.latestudy.model
+
+data class StudyType(
+    val id: String,
+    val name: String,
+)
