+++ data/src/main/kotlin/team/aliens/dms/android/data/latestudy/repository/LateStudyRepository.kt
@@ -1,6 +1,6 @@
-package team.aliens.dms.android.domain.latestudy.repository
+package team.aliens.dms.android.data.latestudy.repository
 
-import team.aliens.dms.android.domain.latestudy.model.StudyType
+import team.aliens.dms.android.data.latestudy.model.StudyType
 
 interface LateStudyRepository {
     suspend fun fetchStudyTypes(): List<StudyType>
